BIG changes, trying to make a footer component which holds the footer. Further splitting up will happen later, lets keep it working.
Stats controller is gone, will need to split up the stats output in several components and a service or similar I guess... (both components need the same calculation-base).
Profile Ctrl can also be changed later into a component. Also there we need the USER in different places, so its rather something else what we can inject everywhere.
Directive?Factory?Service?

refs #26
