+++ js/cyclone.js
@@ -10,6 +10,6 @@ require('./cyclone.module.js');
 require('./components/auth.factory.js');
 require('./components/focus.directive.js');
 require('./components/order.filter.js');
-require('./user/profile.controller.js');
+require('./footer/footer.component');
 require('./time/time.controller.js');
-require('./stats/stats.controller.js');
+require('./version/version.component.js');
