+++ public/index.html
@@ -428,87 +428,7 @@
         <pre>{{ profile | json }}</pre> -->
     </div>
 
-    <footer class="footer-stats"
-            ng-controller="StatsCtrl"
-            ng-cloak>
-        <md-toolbar class="footer-stats__totals md-toolbar-tools"
-                    layout="column"
-                    layout-gt-sm="row"
-                    layout-align="space-between center"
-        >
-            <div flex="nogrow">
-                <div ng-cloak ng-controller="ProfileCtrl" class="card-content white-text"
-                     layout="row" layout-align="start center">
-                    <a ng-click="ProfileOpen = !ProfileOpen" class="profileLink">
-                        <md-icon>account_circle</md-icon>
-                        <span>{{ user }}</span>
-                    </a>
-                    <span ng-if="!user" class="status">{{ connection }}</span>
-                    <div ng-show="ProfileOpen || !user" id="ProfileOpen" layout="row">
-                        <a id="logoutButton"
-                           ng-click="logout()"
-                           class=""
-                           ng-show="user">
-                            <md-button class="md-raised md-primary" aria-label="Login">Logout</md-button>
-                        </a>
-                        <div id="account-details" ng-show="userIsAdmin"></div>
-                    </div>
-
-                    <!--Data: <pre>{{ profile | json }}</pre>-->
-                </div>
-            </div>
-            <stats-total
-                work-total="statsTotalWork"
-                private-total="statsTotalPrivate"
-            ></stats-total>
-            <version-display></version-display>
-        </md-toolbar>
-        <md-toolbar
-                ng-show="user"
-                class="container footer-stats__details md-toolbar-tools"
-                layout="column" layout-align="center stretch">
-
-            <div class="stats-title">
-                <strong>Tasks</strong>
-            </div>
-            <div class="day-visualize">
-                <md-button ng-repeat="entry in refDayVisArray | orderObjectBy:'order':true"
-                           class="dayVisualize-entry project--{{entry.project}}"
-                           ng-class="[
-                    {true: 'checked--yes', false: 'checked--no'}[entry.checked == true],
-                    {true: 'work', false: 'private'}[entry.type == 'work'],
-                 ]"
-                           ng-style="{'background-color':'' + entry._color + '', 'flex-grow':'' + entry._width + ''}"
-                           aria-label="info">
-                    <md-tooltip md-direction="top">
-                        {{entry.project}}: {{entry.text}}
-                    </md-tooltip>
-                    <div class="dayVisualize-entry-content dayVisualize-entry-content--text">{{entry.text}}</div>
-                    <div class="dayVisualize-entry-content dayVisualize-entry-content--duration">{{entry.timestampDuration | date : "H'h' m'min'" : '+0000'}}</div>
-                    <div class="dayVisualize-entry-content dayVisualize-entry-content--time-start">{{entry.timestampStart | date:'H:mm'}}</div>
-                    <div ng-if=$last class="dayVisualize-entry-content dayVisualize-entry-content--time-end">{{entry.timestamp | date:'H:mm'}}</div>
-                </md-button>
-            </div>
-
-            <div class="stats-title">
-                <strong>Project total</strong>
-            </div>
-            <div class="day-visualize project-sum">
-                <md-button ng-repeat="entry in dayVisualizeProjectTotals track by $index"
-                           class="dayVisualize-entry project--{{entry.project}}"
-                           ng-class="[
-                    {true: 'work', false: 'private'}[entry.type == 'work'],
-                 ]"
-                           ng-style="{'background-color':'' + entry._color + '', 'flex-grow':'' + entry._width + ''}"
-                           aria-label="info">
-                    </md-tooltip>
-                    <div class="dayVisualize-entry-content dayVisualize-entry-content--project">{{entry.project}}</div>
-                    <div class="dayVisualize-entry-content dayVisualize-entry-content--time">{{entry.duration | date : "H'h' m'min'" : '+0000'}}</div>
-                </md-button>
-            </div>
-
-        </md-toolbar>
-    </footer>
+    <footer-display class="footer-stats" ng-cloak></footer-display>
 
 </div>
